Refactors accordion
Passing the expander as a slot didn't provide enough flexibility.
In my case, I want to position the expander, more that just left and
right align it compared to the Summary. Instead of using a slot,
a data-handle is used to get the on click functionality. The only
thing we drop is the default styling cursor: pointer which is
very simple to just add (and you might not always want a pointer...)